Yet another attempt to fix the MathJax CDN issue
I changed the CDN script tag to what was specified by the MathJax
documentation. Previuosly, I just chose one randomly, which is not the
best idea. I frequently run into the issue of Mathjax working locally,
but not when I push to Github, so this may not be my last attempt at
this.