Start of complete site redesign
I am now working on a complete redesign of the site. Looking through my
old code, I belive I 'bloated' my code with too many container divs.
I also did not understand a lot of the fundamentals of how to align and
space elements with css. Finally, the site was not responsive.

This redesign has the goal of fixing all of these things:

1. Simple code for a simple project
2. More intelligient use of CSS for spacing and alignment
3. A responsive site

I do not know how long it will take, but I hope to have it finished by
the time I start graduate school.