Sticky footer using calc
The source I used for this:

https://css-tricks.com/couple-takes-sticky-footer/

Basically, I added ONE content wrapper and made its height 100vh - 80px.
THis corresponds to the FIXED  height of my footer (60px) and 20 px of
padding.

That padding was added as padding-top in the footer tag...

This may cause problems in the future? Idk yet.