Changed line height for all non-heading text, major 3rd scaling
The line height is 1.35x the font size.
Standard font size is 16 px with major 3rd scaling all the way from h1
to h5.
H1 may be too big to be useable as a title?
Also, I need to figure out how to make the content, responsive,
especially the nav bar spreading.