Successfully got d3 to work on the page. Solution below:
THe newest version of Hugo uses Goldmark as the default library for
Markdown in Hugo. It is something calles 'CommonMark compliant'. I do
not know what it means. But, I DO know that in this webpage:

https://gohugo.io/news/0.60.0-relnotes/

they tell you:

"Also, if you have lots of inline HTML in your Markdown files, you may
have to enable the unsafe mode:"

this is what the

[markup.goldmark.renderer]
unsafe = true

or similar line is doing. This enables unsafe mode, which is what you
needed.